SLAM_HW4_ICP/code/pyproject.toml
[project]
name = "slam-hw-icp"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "open3d",
    "matplotlib",
    "Pillow",
    "rerun-sdk",
    "tyro",
    "tqdm",
    "numpy-quaternion",
]
